Rearrange CI scripts for cleaner organisation. We are also adding the necessary licensing and rustfmt checks to ensure that the licensing information is present.
To allow us to get cargo-deny we had to add libc6 and make as
dependencies.

Since we are editing the files to add the license headers we're also
moving the mod.rs files to their named equivalents to be in line with
Rust 2018 conventions.

During an update of git-platinum some tests needed to be fixed as well.