Use nom for parsing references (#161)
* Use nom for parsing references

The regex had a subtle bug in it that I wasn't willing to put the effort
in to figure out why. Instead I've switched to use the nom parsing
library so I can have a better understanding of the semantics. However,
I also did not put in the effort to figure out how to make the errors
better. Any failure is considered a "MalformedRef".

* Remove println!

* v0.5.1