Fix Infra (#164)
* Rework adding remote branches

Extending the "heelflip" branch causes issues on CI, since the name can
already exist. It could be that we got away with it due to hitting a
different machine or something the first time.

Instead, we add a new case to test the particular scenario.

* Teardown test branches

To ensure that we have the correct set up for mock branches we teardown
the branches by deleting them. The build.rs file will then create the
correct set.