data/git-platinum: use tarball file to keep track of git-platinum
Instead of using a git submodule to keep track of git-platinum, we use a
cloned version of git-platinum that is compressed into a tarball file.
This gets uncompressed via the `build.rs` file so that it can used by
the tests. A helper script for cloning and compressing git-platinum is
provided to get new updates, while also adding custom branches for our
tests.

Signed-off-by: Tomasz Kurcz <uint@lavabit.com>
Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>